app/requirements.txt
# .app/requirements.txt

alembic==1.13.2
APScheduler==3.10.4
Authlib==1.3.2
bcrypt==4.2.0
fastapi[standard]==0.115.0
itsdangerous==2.2.0
mailjet_rest==1.3.4
pillow==10.4.0
pydantic-settings==2.5.2
pyotp==2.9.0
qrcode==7.4.2
SQLAlchemy==2.0.35
